refactor: split investigation walk reports
