+++ docs/agent-maintainer-hardening-branch-roadmap.md
@@ -1,6 +1,6 @@
 # Agent Maintainer Hardening Branch Roadmap
 
-Current chunk: `refactor/report-action-brief`
+Current chunk: `refactor/report-walk-export`
 
 ## Goal
 
@@ -125,8 +125,8 @@ module boundary and keep behavior covered by focused tests.
     dedicated report family below both line thresholds.
   - [x] Extract action-brief construction and split goal/finding strategy from
     agentic report assembly so both modules remain under threshold.
-  - [ ] Extract investigation, hypothesis, walk/export, and query/recommendation
-    families until the facade is below the configured thresholds.
+  - [x] Extract investigation walk and strict evidence export; the remaining
+    report facade is now below both configured line thresholds.
 - Split `src/codex_usage_tracker/store/content_index.py` by ingestion, FTS,
   provenance, and query responsibilities.
 - Then address medium oversized modules surfaced by the latest file-length log:
